+++ out/bin/compilers/7800basic/7800bas.c
@@ -13,7 +13,7 @@ FILE *stdoutfilepointer;
 
 extern int currentdmahole;
 
-#define BASIC_VERSION_INFO "7800basic v0.16"
+#define BASIC_VERSION_INFO "7800basic v0.17"
 
 int main(int argc, char *argv[])
 {
@@ -66,6 +66,7 @@ int main(int argc, char *argv[])
     bankcount = 0;
     currentbank = 0;
     doublebufferused = 0;
+    boxcollisionused = 0;
     dmaplain = 0;
     templabel = 0;
     doublewide = 0;